Fabrice KAUFFMANN:
I have recently modeled a French Mirage 2000C/RDI fighter with its 2D
cockpit. I still need to adapt the flight model in order to match the real
aircraft data and a lot of improvements can be done, but the current version
is already flyable.

The latest release can be downloaded from:

    http://fkauffmann.ifrance.com/archives

You can also find more of pictures of the Mirage 2000 on my blog:

    http://scifi.skynetblogs.be (sorry, it's in French language)

If the quality of my work is sufficient, I would be very happy to see my
model included in the FlightGear package. Otherwise, let me know what has
to be improved.

Thanks in advance for your answer and congratulations for your outstanding
simulator !

Best regards and Happy New Year,